search: add generic search stream handlers for clients
