+++ cmd/main.go
@@ -7,9 +7,12 @@ import (
 	"io"
 	"net/http"
 	"net/url"
+	"os"
+	"os/signal"
 	"strings"
 
 	"github.com/fatih/color"
+	"github.com/textileio/textile-go/pb"
 	"github.com/textileio/textile-go/util"
 )
 
@@ -71,10 +74,12 @@ func executeStringCmd(meth method, pth string, pars params) (string, error) {
 		return "", err
 	}
 	defer res.Body.Close()
+
 	body, err := util.UnmarshalString(res.Body)
 	if err != nil {
 		return "", err
 	}
+
 	return body, nil
 }
 
@@ -84,6 +89,7 @@ func executeJsonCmd(meth method, pth string, pars params, target interface{}) (s
 		return "", err
 	}
 	defer res.Body.Close()
+
 	if res.StatusCode >= 400 {
 		body, err := util.UnmarshalString(res.Body)
 		if err != nil {
@@ -91,13 +97,16 @@ func executeJsonCmd(meth method, pth string, pars params, target interface{}) (s
 		}
 		return "", errors.New(body)
 	}
+
 	if err := util.UnmarshalJSON(res.Body, target); err != nil {
 		return "", err
 	}
+
 	data, err := json.MarshalIndent(target, "", "    ")
 	if err != nil {
 		return "", err
 	}
+
 	return string(data), nil
 }
 
@@ -107,6 +116,7 @@ func request(meth method, pth string, pars params) (*http.Response, func(), erro
 	if err != nil {
 		return nil, nil, err
 	}
+
 	if len(pars.args) > 0 {
 		var args []string
 		for _, arg := range pars.args {
@@ -114,6 +124,7 @@ func request(meth method, pth string, pars params) (*http.Response, func(), erro
 		}
 		req.Header.Set("X-Textile-Args", strings.Join(args, ","))
 	}
+
 	if len(pars.opts) > 0 {
 		var items []string
 		for k, v := range pars.opts {
@@ -121,16 +132,88 @@ func request(meth method, pth string, pars params) (*http.Response, func(), erro
 		}
 		req.Header.Set("X-Textile-Opts", strings.Join(items, ","))
 	}
+
 	if pars.ctype != "" {
 		req.Header.Set("Content-Type", pars.ctype)
 	}
+
 	tr := &http.Transport{}
 	client := &http.Client{Transport: tr}
 	res, err := client.Do(req)
 	cancel := func() {
 		tr.CancelRequest(req)
 	}
+
 	return res, cancel, err
+}
+
+func handleSearchStream(pth string, param params) {
+	outputCh := make(chan interface{})
+
+	cancel := func() {}
+	quit := make(chan os.Signal)
+	signal.Notify(quit, os.Interrupt)
+
+	go func() {
+		defer func() {
+			cancel()
+			os.Exit(1)
+		}()
+
+		var res *http.Response
+		var err error
+		res, cancel, err = request(POST, pth, param)
+		if err != nil {
+			outputCh <- err.Error()
+			return
+		}
+		defer res.Body.Close()
+
+		if res.StatusCode >= 400 {
+			body, err := util.UnmarshalString(res.Body)
+			if err != nil {
+				outputCh <- err.Error()
+			} else {
+				outputCh <- body
+			}
+			return
+		}
+
+		decoder := json.NewDecoder(res.Body)
+		for decoder.More() {
+			var result *pb.QueryResult
+			if err := decoder.Decode(&result); err == io.EOF {
+				return
+			} else if err != nil {
+				outputCh <- err.Error()
+				return
+			}
+
+			data, err := contactsMarshaler.MarshalToString(result)
+			if err != nil {
+				outputCh <- err.Error()
+				return
+			}
+			outputCh <- data
+		}
+	}()
+
+	for {
+		select {
+		case val := <-outputCh:
+			output(val)
+
+		case <-quit:
+			fmt.Println("Interrupted")
+			if cancel != nil {
+				fmt.Printf("Canceling...")
+				cancel()
+			}
+			fmt.Print("done\n")
+			os.Exit(1)
+			return
+		}
+	}
 }
 
 func output(value interface{}) {
